Code cleanup. Minor fixes.
