--- app/src/main/assets/comments.json
@@ -1,55 +1,0 @@
-{
-  "comments":[
-    {
-      "id": "1",
-      "user": "Joshua",
-      "content": "Hello world",
-      "date": "23/02/2020",
-      "likes": "50",
-      "dislikes": "51",
-      "replies": [
-        {
-          "id": "3",
-          "user": "Lance",
-          "content": "Anta Baka",
-          "date": "23/02/2020",
-          "likes": "50",
-          "dislikes": "51"
-        },
-        {
-          "id": "5",
-          "user": "Lance",
-          "content": "Are you done with the layouts?",
-          "date": "25/02/2020",
-          "likes": "50",
-          "dislikes": "51"
-        }
-      ]
-    },
-    {
-      "id": "2",
-      "user": "Chris",
-      "content": "Aniki...",
-      "date": "23/02/2020",
-      "likes": "50",
-      "dislikes": "51",
-      "replies": [
-        {
-
-        }
-      ]
-    },
-    {
-      "id": "4",
-      "user": "Sheryl",
-      "content": "Shouto danshi!",
-      "date": "23/02/2020",
-      "likes": "50",
-      "dislikes": "51",
-      "replies": [
-        {
-        }
-      ]
-    }
-  ]
-}
