--- app/src/main/java/com/welearn/wemath/QuizMainFragment.java
@@ -1,97 +1,0 @@
-package com.welearn.wemath;
-
-import android.content.Context;
-import android.net.Uri;
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-import androidx.navigation.Navigation;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageButton;
-
-
-public class QuizMainFragment extends Fragment {
-
-    private ImageButton mAnswerButton;
-    private ImageButton mCreateButton;
-
-    private OnFragmentInteractionListener mListener;
-
-    public QuizMainFragment() {
-        // Required empty public constructor
-    }
-
-
-    // TODO: Rename and change types and number of parameters
-    public static QuizMainFragment newInstance(String param1, String param2) {
-        QuizMainFragment fragment = new QuizMainFragment();
-        Bundle args = new Bundle();
-
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View root =  inflater.inflate(R.layout.fragment_quiz_main, container, false);
-
-        mAnswerButton = root.findViewById(R.id.quiz_answering_image);
-        mCreateButton = root.findViewById(R.id.quiz_creation_image);
-
-        mAnswerButton.setOnClickListener(Navigation.createNavigateOnClickListener(R.id.action_navigation_quiz_main_to_quizTakingMenuFragment, null));
-        mCreateButton.setOnClickListener(Navigation.createNavigateOnClickListener(R.id.action_navigation_quiz_main_to_userQuizCreationFragment, null));
-
-
-        return root;
-
-    }
-
-    // TODO: Rename method, update argument and hook method into UI event
-    public void onButtonPressed(Uri uri) {
-        if (mListener != null) {
-            mListener.onFragmentInteraction(uri);
-        }
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        /*if (context instanceof OnFragmentInteractionListener) {
-            mListener = (OnFragmentInteractionListener) context;
-        } else {
-            throw new RuntimeException(context.toString()
-                    + " must implement OnFragmentInteractionListener");
-        }*/
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        mListener = null;
-    }
-
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     * <p>
-     * See the Android Training lesson <a href=
-     * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
-     */
-    public interface OnFragmentInteractionListener {
-        // TODO: Update argument type and name
-        void onFragmentInteraction(Uri uri);
-    }
-}
