+++ app/src/main/java/com/welearn/wemath/lessons/LessonActivity.java
@@ -1,23 +1,24 @@
 package com.welearn.wemath.lessons;
 
+/*activity for displaying the lessons*/
+
+
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.View;
 import android.webkit.WebView;
 import android.widget.Button;
 import android.widget.FrameLayout;
-import android.widget.ScrollView;
 import android.widget.Toast;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.widget.NestedScrollView;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentContainerView;
 import androidx.fragment.app.FragmentManager;
 
 import com.welearn.wemath.R;
+import com.welearn.wemath.lessons.comments.CommentsFragment;
 
 public class LessonActivity extends AppCompatActivity {
 
@@ -29,13 +30,11 @@ public class LessonActivity extends AppCompatActivity {
     private Button mNextButton;
     private Button mPrevButton;
     private Button mCommentsButton;
-    private FragmentContainerView mCommentsFrgament;
     private FragmentManager mFragmentManager;
     private Fragment mFragment;
     private WebView mWebView;
     private NestedScrollView mScrollView;
     private FrameLayout mFrameLayout;
-
 
     private String mYear, mSection;
     private int mTopic, mLesson, mMaxLesson;
@@ -59,27 +58,16 @@ public class LessonActivity extends AppCompatActivity {
 
         mWebView.loadUrl("file:///android_asset/" + mSection + mYear + "_" + mTopic + "." + mLesson + ".html");
 
-
-
-
         mNextButton = findViewById(R.id.next_question_button);
-        mNextButton.setOnClickListener(new View.OnClickListener(){
-            public void onClick(View v){
-                navigate(true);
-            }
-        });
+        mNextButton.setOnClickListener(v -> navigate(true));
 
         mPrevButton = findViewById(R.id.prev_lesson_button);
         if (mLesson == 1){
             mPrevButton.setVisibility(View.GONE);
         }
-        mPrevButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if(mLesson != 1) {
-                    navigate(false);
-
-                }
+        mPrevButton.setOnClickListener(view -> {
+            if(mLesson != 1) {
+                navigate(false);
             }
         });
 
@@ -88,22 +76,20 @@ public class LessonActivity extends AppCompatActivity {
         mFragmentManager = getSupportFragmentManager();
         mFragment = mFragmentManager.findFragmentById(R.id.lesson_comment_fragment);
 
-        mCommentsButton.setOnClickListener(new View.OnClickListener(){
-            public void onClick(View v){
-                if (mFrameLayout.getVisibility() == View.GONE) {
-                    mFrameLayout.setVisibility(View.VISIBLE);
-                    if (mFragment == null) { //check if it's not already there in case of an activity reopening
-                        Bundle bundle = new Bundle();
-                        bundle.putString("choice", "comments/" + mSection + "/" + mYear + "/" + mTopic + "/" + mLesson);
-                        mFragment = new CommentsFragment();
-                        mFragment.setArguments(bundle);
-                        mFragmentManager.beginTransaction().add(R.id.lesson_comment_fragment, mFragment).commit(); //chainable methods because they all return a FragTran
-                    }
-                } else {
-                    mFrameLayout.setVisibility(View.GONE);
+        mCommentsButton.setOnClickListener(v -> {
+            if (mFrameLayout.getVisibility() == View.GONE) {
+                mFrameLayout.setVisibility(View.VISIBLE);
+                if (mFragment == null) { //check if it's not already there in case of an activity reopening
+                    Bundle bundle = new Bundle();
+                    bundle.putString("choice", "comments/" + mSection + "/" + mYear + "/" + mTopic + "/" + mLesson);
+                    mFragment = new CommentsFragment();
+                    mFragment.setArguments(bundle);
+                    mFragmentManager.beginTransaction().add(R.id.lesson_comment_fragment, mFragment).commit();
                 }
-
+            } else {
+                mFrameLayout.setVisibility(View.GONE);
             }
+
         });
     }
 
@@ -135,27 +121,22 @@ public class LessonActivity extends AppCompatActivity {
                     mPrevButton.setVisibility(View.VISIBLE);
                 }
                 mWebView.loadUrl("file:///android_asset/" + mSection + mYear + "_" + mTopic + "." + mLesson + ".html");
-                //mWebView.pageUp(true);
                 mScrollView.fullScroll(View.FOCUS_UP);
-                if(mFragmentManager.getBackStackEntryCount() > 0) {
-                    mFragmentManager.beginTransaction().detach(mFragment).commit();
-                    mFragment = null;
-                    mFrameLayout.setVisibility(View.GONE);
-                }
-
-            } else if (resultCode == PREV_REQUEST_CODE){
+            }
+            else if (resultCode == PREV_REQUEST_CODE){
                 if (mLesson != 1){
                     mPrevButton.setVisibility(View.VISIBLE);
                 } else
                     mPrevButton.setVisibility(View.GONE);
                 mWebView.loadUrl("file:///android_asset/" + mSection + mYear + "_" + mTopic + "." + mLesson + ".html");
-                if(mFragmentManager.getBackStackEntryCount() > 0) {
-                    mFragmentManager.beginTransaction().detach(mFragment).commit();
-                    mFragment = null;
-                    mFrameLayout.setVisibility(View.GONE);
-                }
                 mScrollView.fullScroll(View.FOCUS_DOWN);
+            }
 
+            //close the comments fragments if it is open
+            if(mFragmentManager.getFragments().size()> 0) {
+                mFragmentManager.beginTransaction().detach(mFragment).commit();
+                mFragment = null;
+                mFrameLayout.setVisibility(View.GONE);
             }
         }
     }
