+++ app/src/main/java/com/welearn/wemath/lessons/LessonQuestion.java
@@ -12,6 +12,5 @@ public class LessonQuestion extends Question {
 
     LessonQuestion(String problem, ArrayList<Pair<String, Boolean>> answers, boolean multipleChoice){
         super(problem, answers, multipleChoice);
-
     }
 }
