--- app/src/main/java/com/welearn/wemath/lessons/LessonSelectionViewModel.java
@@ -1,9 +1,0 @@
-package com.welearn.wemath.lessons;
-
-import androidx.lifecycle.ViewModel;
-
-/*ignore for now*/
-
-public class LessonSelectionViewModel extends ViewModel {
-    // TODO: Implement the ViewModel
-}
