--- app/src/main/java/com/welearn/wemath/lessons/LessonTopicViewModel.java
@@ -1,27 +1,0 @@
-package com.welearn.wemath.lessons;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-/*ignore for now*/
-
-public class LessonTopicViewModel extends ViewModel {
-
-    private String mYear;
-    private String mSection;
-
-    public LessonTopicViewModel(String year, String section){
-
-        mYear = year;
-        mSection = section;
-    }
-
-    public String getYear() {
-        return mYear;
-    }
-
-    public String getSection() {
-        return mYear;
-    }
-}
