+++ app/src/main/java/com/welearn/wemath/lessons/LessonYearFragment.java
@@ -1,6 +1,9 @@
 package com.welearn.wemath.lessons;
 
+/*fragment for selecting the school year*/
+
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.cardview.widget.CardView;
 import androidx.lifecycle.ViewModelProviders;
 
 import android.os.Bundle;
@@ -14,16 +17,14 @@ import androidx.navigation.Navigation;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageButton;
-import android.widget.TextView;
 
 import com.welearn.wemath.R;
 
 public class LessonYearFragment extends Fragment {
 
-    private LessonYearViewModel mViewModel;
-    private ImageButton mYear1, mYear2, mYear3;
-    private TextView mYear1Text, mYear2Text, mYear3Text;
+    private CardView mYear1;
+    private CardView mYear2;
+    private CardView mYear3;
 
     public static LessonYearFragment newInstance() {
         return new LessonYearFragment();
@@ -40,13 +41,10 @@ public class LessonYearFragment extends Fragment {
         //set the actionbar title
         ((AppCompatActivity) getActivity()).getSupportActionBar().setTitle(section.toUpperCase());
 
-        mYear1 = root.findViewById(R.id.year1_button);
-        mYear2 = root.findViewById(R.id.year2_button);
-        mYear3 = root.findViewById(R.id.year3_button);
+        mYear1 = root.findViewById(R.id.Year1_card);
+        mYear2 = root.findViewById(R.id.Year2_card);
+        mYear3 = root.findViewById(R.id.Year3_card);
 
-        mYear1Text = root.findViewById(R.id.year1_text);
-        mYear2Text = root.findViewById(R.id.year2_text);
-        mYear3Text = root.findViewById(R.id.year3_text);
 
         /*Setting the behaviours of the buttons*/
 
@@ -62,17 +60,6 @@ public class LessonYearFragment extends Fragment {
             }
         });
 
-        mYear1Text.setOnClickListener(new View.OnClickListener(){
-            public void onClick(View v){
-                //parameters to be passed to the topics fragment
-                String year = "1";
-
-                //set up the navigation action with the parameters
-                NavDirections action = LessonYearFragmentDirections.actionLessonYearFragmentToLessonTopicFragment(year, section);
-                Navigation.findNavController(v).navigate(action);
-
-            }
-        });
 
         mYear2.setOnClickListener(new View.OnClickListener(){
             public void onClick(View v){
@@ -84,15 +71,6 @@ public class LessonYearFragment extends Fragment {
             }
         });
 
-        mYear2Text.setOnClickListener(new View.OnClickListener(){
-            public void onClick(View v){
-                String year = "2";
-                //String section = LessonYearFragmentArgs.fromBundle(getArguments()).getSection();
-                NavDirections action = LessonYearFragmentDirections.actionLessonYearFragmentToLessonTopicFragment(year, section);
-                Navigation.findNavController(v).navigate(action);
-
-            }
-        });
 
         mYear3.setOnClickListener(new View.OnClickListener(){
             public void onClick(View v){
@@ -104,24 +82,8 @@ public class LessonYearFragment extends Fragment {
             }
         });
 
-        mYear3Text.setOnClickListener(new View.OnClickListener(){
-            public void onClick(View v){
-                String year = "3";
-                //String section = LessonYearFragmentArgs.fromBundle(getArguments()).getSection();
-                NavDirections action = LessonYearFragmentDirections.actionLessonYearFragmentToLessonTopicFragment(year, section);
-                Navigation.findNavController(v).navigate(action);
-
-            }
-        });
         return root;
     }
 
-    //not used yet
-    @Override
-    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-        mViewModel = ViewModelProviders.of(this).get(LessonYearViewModel.class);
-        // TODO: Use the ViewModel
-    }
 
 }
