--- app/src/main/java/com/welearn/wemath/lessons/LessonYearViewModel.java
@@ -1,9 +1,0 @@
-package com.welearn.wemath.lessons;
-
-import androidx.lifecycle.ViewModel;
-
-/*ignore for now*/
-
-public class LessonYearViewModel extends ViewModel {
-    // TODO: Implement the ViewModel
-}
