+++ app/src/main/java/com/welearn/wemath/login/LoginActivity.java
@@ -1,6 +1,5 @@
-package com.welearn.wemath;
+package com.welearn.wemath.login;
 
-import androidx.activity.OnBackPressedDispatcher;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -29,16 +28,17 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.auth.UserInfo;
 import com.google.firebase.auth.UserProfileChangeRequest;
+import com.welearn.wemath.MainActivity;
+import com.welearn.wemath.R;
 
 
 public class LoginActivity extends AppCompatActivity {
 
     private static final String TAG = "EmailPassword";
 
-
     private FirebaseAuth mAuth;
     private Button mRegisterButton, mSignInButton, mSignOutButton;
-    TextView mAnonymousButton;
+    private TextView mAnonymousButton;
     private EditText mEmail, mPassword;
     private CallbackManager mCallbackManager;
     private LoginButton mFacebookloginButton;
@@ -46,9 +46,9 @@ public class LoginActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        //FacebookSdk.fullyInitialize();
         setContentView(R.layout.activity_login);
-        // Initialize Firebase Auth
+
+
         mAuth = FirebaseAuth.getInstance();
 
         mRegisterButton = findViewById(R.id.login_activity_register_button);
@@ -64,69 +64,55 @@ public class LoginActivity extends AppCompatActivity {
         FirebaseUser currentUser = mAuth.getCurrentUser();
         updateUI(currentUser);
 
-        mRegisterButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick (View v){
-                if (mEmail.getText().toString().trim().length() != 0 && mPassword.getText().toString().trim().length() != 0 ) {
-                    Toast.makeText(LoginActivity.this, "Creating new account...", Toast.LENGTH_SHORT).show();
-                    createUser(mEmail.getText().toString(), mPassword.getText().toString());
-                    mEmail.setText("");
-                    mPassword.setText("");
-                } else{
-                    Toast.makeText(LoginActivity.this, "Please enter a valid email and password", Toast.LENGTH_SHORT).show();
-                }
-
+        mRegisterButton.setOnClickListener(v -> {
+            if (mEmail.getText().toString().trim().length() != 0 && mPassword.getText().toString().trim().length() != 0 ) {
+                Toast.makeText(LoginActivity.this, "Creating new account...", Toast.LENGTH_SHORT).show();
+                createUser(mEmail.getText().toString(), mPassword.getText().toString());
+                mEmail.setText("");
+                mPassword.setText("");
+            } else{
+                Toast.makeText(LoginActivity.this, "Please enter a valid email and password", Toast.LENGTH_SHORT).show();
             }
+
         });
 
-        mSignInButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick (View v){
-                if (mEmail.getText().toString().trim().length() != 0 && mPassword.getText().toString().trim().length() != 0 ) {
-                    Toast.makeText(LoginActivity.this, "Signing in...", Toast.LENGTH_SHORT).show();
-                    signInUser(mEmail.getText().toString(), mPassword.getText().toString());
-                    mEmail.setText("");
-                    mPassword.setText("");
-
-                } else{
-                    Toast.makeText(LoginActivity.this, "Please enter a valid email and password", Toast.LENGTH_SHORT).show();
-                }
+        mSignInButton.setOnClickListener(v -> {
+            if (mEmail.getText().toString().trim().length() != 0 && mPassword.getText().toString().trim().length() != 0 ) {
+                Toast.makeText(LoginActivity.this, "Signing in...", Toast.LENGTH_SHORT).show();
+                signInUser(mEmail.getText().toString(), mPassword.getText().toString());
+                mEmail.setText("");
+                mPassword.setText("");
+            } else{
+                Toast.makeText(LoginActivity.this, "Please enter a valid email and password", Toast.LENGTH_SHORT).show();
             }
         });
 
 
-        mAnonymousButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick (View v){
-                Toast.makeText(LoginActivity.this, "Logging in...", Toast.LENGTH_SHORT).show();
-                signInAnon();
+        mAnonymousButton.setOnClickListener(v -> {
+            Toast.makeText(LoginActivity.this, "Logging in...", Toast.LENGTH_SHORT).show();
+            signInAnon();
 
-            }
         });
 
-        mSignOutButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick (View v){
-                if(mAuth.getCurrentUser().isAnonymous()){
+        mSignOutButton.setOnClickListener(v -> {
+            if(mAuth.getCurrentUser().isAnonymous()){
+                mAuth.signOut();
+                updateUI(null);
+            }else {
+                UserInfo profile = mAuth.getCurrentUser().getProviderData().get(1);
+                if (profile.getProviderId().equalsIgnoreCase("facebook.com")) {
+                    mAuth.signOut();
+                    LoginManager.getInstance().logOut();
+                    updateUI(null);
+                } else {
                     mAuth.signOut();
                     updateUI(null);
-                }else {
-                    UserInfo profile = mAuth.getCurrentUser().getProviderData().get(1);
-                    if (profile.getProviderId().equalsIgnoreCase("facebook.com")) {
-                        mAuth.signOut();
-                        LoginManager.getInstance().logOut();
-                        updateUI(null);
-                    } else {
-                        mAuth.signOut();
-                        updateUI(null);
-                    }
                 }
-                Intent intent = new Intent(getApplicationContext(), LoginActivity.class);
-                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                startActivity(intent);
-                //LoginManager.getInstance().logOut();
-
             }
+            Intent intent = new Intent(getApplicationContext(), LoginActivity.class);
+            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+            startActivity(intent);
+
         });
 
         mCallbackManager = CallbackManager.Factory.create();
@@ -142,38 +128,31 @@ public class LoginActivity extends AppCompatActivity {
             @Override
             public void onCancel() {
                 Log.d(TAG, "facebook:onCancel");
-                // ...
             }
 
             @Override
             public void onError(FacebookException error) {
                 Log.d(TAG, "facebook:onError", error);
-                // ...
             }
         });
-// ...
     }
+
         @Override
         protected void onActivityResult(int requestCode, int resultCode, Intent data) {
             super.onActivityResult(requestCode, resultCode, data);
 
             // Pass the activity result back to the Facebook SDK
             mCallbackManager.onActivityResult(requestCode, resultCode, data);
-
     }
-
 
 
     @Override
     public void onStart() {
         super.onStart();
-
-
     }
 
     public void updateUI(FirebaseUser user){
         if(user != null){
-
                 mEmail.setVisibility(View.GONE);
                 mPassword.setVisibility(View.GONE);
                 mSignInButton.setVisibility(View.GONE);
@@ -183,7 +162,6 @@ public class LoginActivity extends AppCompatActivity {
                 mSignOutButton.setVisibility((View.VISIBLE));
 
         } else{
-
             mSignOutButton.setVisibility((View.GONE));
             mEmail.setVisibility(View.VISIBLE);
             mPassword.setVisibility(View.VISIBLE);
@@ -204,10 +182,12 @@ public class LoginActivity extends AppCompatActivity {
                         if (task.isSuccessful()) {
                             // Sign in success, update UI with the signed-in user's information
                             Log.d(TAG, "createUserWithEmail:success");
+
+                            createPlaceholderName();
+
                             Toast.makeText(LoginActivity.this, "Success!",
                                     Toast.LENGTH_SHORT).show();
-                            FirebaseUser user = mAuth.getCurrentUser();
-                            updateUI(user);
+
                             Intent intent = new Intent(getApplicationContext(), LoginRegisterActivity.class);
                             startActivity(intent);
                             finish();
@@ -218,8 +198,19 @@ public class LoginActivity extends AppCompatActivity {
                                     Toast.LENGTH_SHORT).show();
                             updateUI(null);
                         }
+                    }
+                });
+    }
 
-                        // ...
+    public void createPlaceholderName(){
+        UserProfileChangeRequest profileUpdates = new UserProfileChangeRequest.Builder()
+                .setDisplayName("NoName")
+                .build();
+        FirebaseUser user = mAuth.getCurrentUser();
+        user.updateProfile(profileUpdates)
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+                        Log.d("Profile update", "User profile updated.");
                     }
                 });
     }
@@ -234,8 +225,6 @@ public class LoginActivity extends AppCompatActivity {
                             Log.d(TAG, "signInWithEmail:success");
                             Toast.makeText(LoginActivity.this, "Sign-in successful.",
                                     Toast.LENGTH_SHORT).show();
-                            FirebaseUser user = mAuth.getCurrentUser();
-                            updateUI(user);
                             redirect();
                         } else {
                             // If sign in fails, display a message to the user.
@@ -243,8 +232,6 @@ public class LoginActivity extends AppCompatActivity {
                             Toast.makeText(LoginActivity.this, "Sign-in failed.", Toast.LENGTH_SHORT).show();
                             updateUI(null);
                         }
-
-                        // ...
                     }
                 });
     }
@@ -257,9 +244,7 @@ public class LoginActivity extends AppCompatActivity {
                         if (task.isSuccessful()) {
                             // Sign in success, update UI with the signed-in user's information
                             Log.d(TAG, "signInAnonymously:success");
-                            //FirebaseUser user = mAuth.getCurrentUser();
                             createAnonUser();
-                            //redirect();
                         } else {
                             // If sign in fails, display a message to the user.
                             Log.w(TAG, "signInAnonymously:failure", task.getException());
@@ -267,8 +252,6 @@ public class LoginActivity extends AppCompatActivity {
                                     Toast.LENGTH_SHORT).show();
                             updateUI(null);
                         }
-
-                        // ...
                     }
                 });
 
@@ -286,13 +269,10 @@ public class LoginActivity extends AppCompatActivity {
                         if (task.isSuccessful()) {
                             Log.d("Profile update", "User profile updated.");
                             Toast.makeText(getBaseContext(), "Logged-in as Guest", Toast.LENGTH_SHORT).show();
-                            updateUI(mAuth.getCurrentUser());
                             redirect();
                         }
                     }
                 });
-
-
     }
 
     private void handleFacebookAccessToken(AccessToken token) {
@@ -319,11 +299,9 @@ public class LoginActivity extends AppCompatActivity {
                             updateUI(null);
                         }
 
-                        // ...
                     }
                 });
     }
-
 
     public void redirect(){
         Intent intent = new Intent(getApplicationContext(), MainActivity.class);
