+++ app/src/main/java/com/welearn/wemath/quizzes/Admin/AdminQuiz.java
@@ -1,0 +1,6 @@
+package com.welearn.wemath.quizzes.Admin;
+
+import com.welearn.wemath.quizzes.Quiz;
+
+class AdminQuiz extends Quiz {
+}
