--- app/src/main/java/com/welearn/wemath/quizzes/AdminQuiz.java
@@ -1,6 +1,0 @@
-package com.welearn.wemath.quizzes;
-
-import com.welearn.wemath.Quiz;
-
-class AdminQuiz extends Quiz {
-}
