+++ app/src/main/res/layout/activity_login.xml
@@ -9,7 +9,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".LoginActivity">
+    tools:context=".login.LoginActivity">
 
     <EditText
         android:id="@+id/login_activity_email"
