+++ app/src/main/res/layout/fragment_user_quiz_creation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".quizzes.UserQuizCreationFragment">
+    tools:context=".quizzes.User.UserQuizCreationFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
