+++ app/src/main/res/layout/fragment_user_quiz_creation_question.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".quizzes.UserQuizCreationQuestionFragment">
+    tools:context=".quizzes.User.UserQuizCreationQuestionFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
