+++ app/src/main/res/layout/fragment_user_quiz_selection.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".quizzes.UserQuizSelectionFragment"
+    tools:context=".quizzes.User.UserQuizSelectionFragment"
     android:fillViewport="true">
 
     <LinearLayout
