+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/navigation_lesson_section"
-        android:name="com.welearn.wemath.LessonSectionFragment"
+        android:name="com.welearn.wemath.lessons.LessonSectionFragment"
         android:label="Section Selection"
         tools:layout="@layout/fragment_lesson_section">
         <action
@@ -21,7 +21,7 @@
         tools:layout="@layout/fragment_home_screen" />
     <fragment
         android:id="@+id/navigation_quiz_main"
-        android:name="com.welearn.wemath.QuizMainFragment"
+        android:name="com.welearn.wemath.quizzes.QuizMainFragment"
         android:label="Quiz Home"
         tools:layout="@layout/fragment_quiz_main">
         <action
@@ -99,7 +99,7 @@
     </activity>
     <fragment
         android:id="@+id/adminQuizSelection"
-        android:name="com.welearn.wemath.quizzes.AdminQuizSelectionFragment"
+        android:name="com.welearn.wemath.quizzes.Admin.AdminQuizSelectionFragment"
         android:label="Admin Quiz"
         tools:layout="@layout/fragment_admin_quiz_selection">
         <action
@@ -108,7 +108,7 @@
     </fragment>
     <activity
         android:id="@+id/activityuserquiztaking"
-        android:name="com.welearn.wemath.quizzes.UserQuizTakingActivity"
+        android:name="com.welearn.wemath.quizzes.User.UserQuizTakingActivity"
         android:label="User Quiz"
         tools:layout="@layout/activity_user_quiz_taking">
         <argument
@@ -132,7 +132,7 @@
     </fragment>
     <fragment
         android:id="@+id/userQuizSelection"
-        android:name="com.welearn.wemath.quizzes.UserQuizSelectionFragment"
+        android:name="com.welearn.wemath.quizzes.User.UserQuizSelectionFragment"
         android:label="User Quiz"
         tools:layout="@layout/fragment_user_quiz_selection">
         <action
@@ -141,7 +141,7 @@
     </fragment>
     <fragment
         android:id="@+id/userQuizCreationFragment"
-        android:name="com.welearn.wemath.quizzes.UserQuizCreationFragment"
+        android:name="com.welearn.wemath.quizzes.User.UserQuizCreationFragment"
         android:label="Quiz Creation"
         tools:layout="@layout/fragment_user_quiz_creation">
         <action
@@ -150,7 +150,7 @@
     </fragment>
     <fragment
         android:id="@+id/userQuizCreationQuestionFragment"
-        android:name="com.welearn.wemath.quizzes.UserQuizCreationQuestionFragment"
+        android:name="com.welearn.wemath.quizzes.User.UserQuizCreationQuestionFragment"
         android:label="Question Creation"
         tools:layout="@layout/fragment_user_quiz_creation_question">
         <argument
